+++ src/lib/index.d.ts
@@ -10,6 +10,7 @@ export type BookingSelectorSelectionScheme = 'linear' | 'square'
 
 export interface BookingSelectorProps {
   className?: string
+  id?: string
   selection?: BookingSelectorDateList
   blocked?: BookingSelectorDateList
   selectionScheme?: BookingSelectorSelectionScheme